convert to initial migration
