--- backend/alembic/versions/041e7e094383_add_timeout_monitor_and_cleanup_timeout_.py
@@ -1,35 +1,0 @@
-"""add_timeout_monitor_and_cleanup_timeout_tasks_enum_values
-
-Revision ID: 041e7e094383
-Revises: 2a6c3b7ee10f
-Create Date: 2025-08-17 17:22:08.866442
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '041e7e094383'
-down_revision: Union[str, None] = '2a6c3b7ee10f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # Add new enum values to tasktype enum
-    op.execute("ALTER TYPE tasktype ADD VALUE 'TIMEOUT_MONITOR'")
-    op.execute("ALTER TYPE tasktype ADD VALUE 'CLEANUP_TIMEOUT_TASKS'")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    # Note: PostgreSQL doesn't support removing enum values directly
-    # This would require recreating the enum and updating all dependent tables
-    # For production, consider a more complex downgrade strategy
-    pass
-    # ### end Alembic commands ###
