--- backend/alembic/versions/7383ea140bc9_make_config_id_nullable_in_task_.py
@@ -1,34 +1,0 @@
-"""Make config_id nullable in task_executions
-
-Revision ID: 7383ea140bc9
-Revises: 62a69389bfb6
-Create Date: 2025-08-15 19:56:34.239086
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '7383ea140bc9'
-down_revision: Union[str, None] = '62a69389bfb6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('task_executions', 'config_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('task_executions', 'config_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    # ### end Alembic commands ###
