--- backend/alembic/versions/cede1b500688_fix_cascade_delete_for_schedule_events_.py
@@ -1,36 +1,0 @@
-"""fix_cascade_delete_for_schedule_events_and_task_executions
-
-Revision ID: cede1b500688
-Revises: 4a4166c0b141
-Create Date: 2025-08-17 06:36:19.554706
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'cede1b500688'
-down_revision: Union[str, None] = '4a4166c0b141'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f('schedule_events_config_id_fkey'), 'schedule_events', type_='foreignkey')
-    op.create_foreign_key(None, 'schedule_events', 'task_configs', ['config_id'], ['id'], ondelete='CASCADE')
-    op.drop_constraint(op.f('task_executions_config_id_fkey'), 'task_executions', type_='foreignkey')
-    op.create_foreign_key(None, 'task_executions', 'task_configs', ['config_id'], ['id'], ondelete='SET NULL')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'task_executions', type_='foreignkey')
-    op.create_foreign_key(op.f('task_executions_config_id_fkey'), 'task_executions', 'task_configs', ['config_id'], ['id'])
-    op.drop_constraint(None, 'schedule_events', type_='foreignkey')
-    op.create_foreign_key(op.f('schedule_events_config_id_fkey'), 'schedule_events', 'task_configs', ['config_id'], ['id'])
-    # ### end Alembic commands ###
