--- backend/alembic/versions/d80b093e0800_remove_test_auto_enum.py
@@ -1,30 +1,0 @@
-"""remove_test_auto_enum
-
-Revision ID: d80b093e0800
-Revises: 8a6729288167
-Create Date: 2025-08-17 17:47:58.286906
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd80b093e0800'
-down_revision: Union[str, None] = '8a6729288167'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values('public', 'tasktype', ['BATCH_SCRAPING', 'BOT_SCRAPING', 'CLEANUP_CONTENT', 'CLEANUP_EVENTS', 'CLEANUP_TIMEOUT_TASKS', 'CLEANUP_TOKENS', 'DATA_ANALYSIS', 'DATA_BACKUP', 'DATA_EXPORT', 'HEALTH_CHECK', 'LOG_ROTATION', 'MANUAL_SCRAPING', 'SEND_EMAIL', 'SEND_NOTIFICATION', 'SYSTEM_MONITOR', 'TEST_AUTO_ENUM', 'TIMEOUT_MONITOR'], ['BATCH_SCRAPING', 'BOT_SCRAPING', 'CLEANUP_CONTENT', 'CLEANUP_EVENTS', 'CLEANUP_TIMEOUT_TASKS', 'CLEANUP_TOKENS', 'DATA_ANALYSIS', 'DATA_BACKUP', 'DATA_EXPORT', 'HEALTH_CHECK', 'LOG_ROTATION', 'MANUAL_SCRAPING', 'SEND_EMAIL', 'SEND_NOTIFICATION', 'SYSTEM_MONITOR', 'TIMEOUT_MONITOR'], [('task_configs', 'task_type')], False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values('public', 'tasktype', ['BATCH_SCRAPING', 'BOT_SCRAPING', 'CLEANUP_CONTENT', 'CLEANUP_EVENTS', 'CLEANUP_TIMEOUT_TASKS', 'CLEANUP_TOKENS', 'DATA_ANALYSIS', 'DATA_BACKUP', 'DATA_EXPORT', 'HEALTH_CHECK', 'LOG_ROTATION', 'MANUAL_SCRAPING', 'SEND_EMAIL', 'SEND_NOTIFICATION', 'SYSTEM_MONITOR', 'TIMEOUT_MONITOR'], ['BATCH_SCRAPING', 'BOT_SCRAPING', 'CLEANUP_CONTENT', 'CLEANUP_EVENTS', 'CLEANUP_TIMEOUT_TASKS', 'CLEANUP_TOKENS', 'DATA_ANALYSIS', 'DATA_BACKUP', 'DATA_EXPORT', 'HEALTH_CHECK', 'LOG_ROTATION', 'MANUAL_SCRAPING', 'SEND_EMAIL', 'SEND_NOTIFICATION', 'SYSTEM_MONITOR', 'TEST_AUTO_ENUM', 'TIMEOUT_MONITOR'], [('task_configs', 'task_type')], True)
-    # ### end Alembic commands ###
